Set default for Process.get to empty bitstring
It is not always the case that the process that `use`s Puid is the one to call `generate/0`. When this happens, the `@puid_carried_bits` key in the process dictionary is not going to be set, and we end up calling `bit_size(nil)`, which crashes. Instead of trying to set the key when the module is built, just use `<<>>` as the default value, which works for every process.